@@ -1,9 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
     <PropertyGroup>
         <Description>Entity Framework Core Second Level Caching Library.</Description>
-        <VersionPrefix>3.8.3</VersionPrefix>
+        <VersionPrefix>3.8.5</VersionPrefix>
         <Authors>Vahid Nasiri</Authors>
-        <TargetFrameworks>net7.0;net6.0;net5.0;netstandard2.1;netstandard2.0;net462;netcoreapp3.1;</TargetFrameworks>
+        <TargetFrameworks>net8.0;net7.0;net6.0;net5.0;netstandard2.1;netstandard2.0;net462;netcoreapp3.1;</TargetFrameworks>
         <GenerateDocumentationFile>true</GenerateDocumentationFile>
         <AssemblyName>EFCoreSecondLevelCacheInterceptor</AssemblyName>
         <PackageId>EFCoreSecondLevelCacheInterceptor</PackageId>
